+++ packages/popover/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @siakit/popover
 
+## 0.0.5
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @siakit/core@0.0.18
+
+## 0.0.4
+
+### Patch Changes
+
+- fix date picker zindex
+
 ## 0.0.3
 
 ### Patch Changes